requirements.txt
chainlit
python-dotenv
openai
pandas
tabulate
asyncpg
numpy
httpx
requests
snowflake-connector-python[pandas]
psycopg2-binary
matplotlib
plotly
seaborn
azure-identity
azure-ai-projects
azure-cosmos
uuid